// src/app/generated/player-api/model/systemPermission.ts
/**
 * Player API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type SystemPermission = 'CreateViews' | 'ViewViews' | 'EditViews' | 'ManageViews' | 'ViewUsers' | 'ManageUsers' | 'ViewApplications' | 'ManageApplications' | 'ViewRoles' | 'ManageRoles' | 'ViewWebhookSubscriptions' | 'ManageWebhookSubscriptions';

export const SystemPermission = {
    CreateViews: 'CreateViews' as SystemPermission,
    ViewViews: 'ViewViews' as SystemPermission,
    EditViews: 'EditViews' as SystemPermission,
    ManageViews: 'ManageViews' as SystemPermission,
    ViewUsers: 'ViewUsers' as SystemPermission,
    ManageUsers: 'ManageUsers' as SystemPermission,
    ViewApplications: 'ViewApplications' as SystemPermission,
    ManageApplications: 'ManageApplications' as SystemPermission,
    ViewRoles: 'ViewRoles' as SystemPermission,
    ManageRoles: 'ManageRoles' as SystemPermission,
    ViewWebhookSubscriptions: 'ViewWebhookSubscriptions' as SystemPermission,
    ManageWebhookSubscriptions: 'ManageWebhookSubscriptions' as SystemPermission
};